Updated a number of things, see description.
Updated our Domain to match the class diagram, which was updated to match Lars' specification and idees, and also include the introduction of the playlist that the domain model showed.

To this i also made a few changes to the Class Diagram, to fix minor oversights and errors.

Finally i also updated our Entity Framework config files to function with the new changes, along with adding some required uniqueness to our domain class via the entity framework